Round 1 clean up complete (and changed file name).
